requirements.txt
fastapi
uvicorn
SQLAlchemy
passlib
bcrypt
python-jose
python-multipart
email-validator
pydantic[dotenv,email]
arq
alembic
aiosmtplib
python-dotenv
asyncpg
greenlet
pyhumps
uvloop
gunicorn
pytest
pytest-asyncio
pytest-cov
mypy
httpx
asgi-lifespan
